Add files via upload
Fraud detection for credit card is a common methodology applied by financial institations in the globe. As a person having strong interest in FinTech and risk-management, it's great to apply a seriea of machine learning models to analyse and make fraud predictions based on the credit card transaction data in September 2013 by european cardholders. It is also a great opportunity for me to learn the magic of conducting machine learning using Scikit-learn.